Refactor authentication module and update dependencies
Removed internal auth guard and shifted authentication responsibilities to sensor.service. Updated all associated file references in other modules. Also made significant changes in package dependencies, most notably in passport.js related modules. This is to improve the overall architecture of the authentication structure and to update the necessary libraries used in the project. It allows for a cleaner and clearer codebase and keeps dependencies up-to-date.