Move condition check to workflows from jobs in CircleCI config
The condition that was previously on the jobs level is transferred to the workflows level in the .circleci/config.yml file. This refactoring can provide a clearer view on when workflows should be triggered according to pipeline.parameters.GHA_Action, and potentially provide a cleaner and more efficient pipeline structure.